chore: some refactor
